\name{BLASTnParse-package}
\alias{BLASTnParse-package}
\alias{BLASTnParse}
\docType{package}
\title{
  Functions to automate several BLAST sites
}
\description{
  Will contain functions to automate BLASTs
}
\details{
\tabular{ll}{
Package: \tab BLASTnParse\cr
Type: \tab Package\cr
Version: \tab 0.0.1-1\cr
Date: \tab 2012-03-06\cr
License: \tab GPL (>= 3)\cr
}
}
\author{
Author: Gerben Gaastra
Maintainer: Gerben Gaastra <gaastra@gtga.nl>
}
\references{
www.rug.nl/gbic
}
\keyword{ package }
\examples{
#TODO
}
